+++ packages/schema-type/src/typescript/index.ts
@@ -1,1 +1,1 @@
-// here!
+export * from "./schemaAbiTypes";
